Refactor Script into multiple files. Break apart spec as well. #114
