+++ spec/script/scripting_spec.rb
@@ -1,0 +1,299 @@
+# encoding: ascii-8bit
+
+# Copyright 2014 Ball Aerospace & Technologies Corp.
+# All Rights Reserved.
+#
+# This program is free software; you can modify and/or redistribute it
+# under the terms of the GNU General Public License
+# as published by the Free Software Foundation; version 3 with
+# attribution addendums as found in the LICENSE.txt
+
+require 'spec_helper'
+require 'cosmos'
+require 'cosmos/script/script'
+require 'tempfile'
+
+module Cosmos
+
+  describe Script do
+
+    before(:all) do
+      cts = File.join(Cosmos::USERPATH,'config','tools','cmd_tlm_server','cmd_tlm_server.txt')
+      FileUtils.mkdir_p(File.dirname(cts))
+      File.open(cts,'w') do |file|
+        file.puts 'INTERFACE INST_INT interface.rb'
+        file.puts 'TARGET INST'
+      end
+      System.class_eval('@@instance = nil')
+
+      require 'cosmos/script'
+    end
+
+    after(:all) do
+      clean_config()
+      FileUtils.rm_rf File.join(Cosmos::USERPATH,'config','tools')
+    end
+
+    before(:each) do
+      allow_any_instance_of(Interface).to receive(:connected?).and_return(true)
+      allow_any_instance_of(Interface).to receive(:disconnect)
+      allow_any_instance_of(Interface).to receive(:write)
+      allow_any_instance_of(Interface).to receive(:read)
+
+      @server = CmdTlmServer.new
+      shutdown_cmd_tlm()
+      initialize_script_module()
+      sleep 0.1
+    end
+
+    after(:each) do
+      @server.stop
+      shutdown_cmd_tlm()
+      sleep(0.1)
+    end
+
+    describe "play_wav_file" do
+      it "plays a wav file if Qt is available" do
+        module Qt
+          def self.execute_in_main_thread(bool); yield; end
+          class CoreApplication; def self.instance; true; end; end;
+          class Sound; def self.isAvailable; true; end; end
+        end
+        expect(Qt::Sound).to receive(:play).with("sound.wav")
+        play_wav_file("sound.wav")
+      end
+    end
+
+    describe "status_bar" do
+      it "sets the ScriptRunner status bar" do
+        class ScriptRunner; end
+        sc = ScriptRunner.new
+        expect(sc).to receive(:script_set_status).with("HI")
+        status_bar("HI")
+      end
+    end
+
+    describe "ask_string, ask" do
+      it "gets user input" do
+        $stdout = StringIO.new
+        expect(self).to receive(:gets) { '10' }
+        expect(ask_string("")).to eql '10'
+        expect(self).to receive(:gets) { '10' }
+        expect(ask("")).to eql 10
+        $stdout = STDOUT
+      end
+    end
+
+    describe "prompt, prompt_message_box" do
+      it "prompts the user for input" do
+        $stdout = StringIO.new
+        expect(self).to receive(:gets) { 'message' }
+        expect(prompt("")).to eql 'message'
+        expect(self).to receive(:gets) { 'b1' }
+        expect(message_box("",["b1","b2"])).to eql 'b1'
+        $stdout = STDOUT
+      end
+    end
+
+    describe "check, check_formatted, check_with_units, check_raw" do
+      it "checks a telemetry item vs a condition" do
+        capture_io do |stdout|
+          check("INST HEALTH_STATUS TEMP1 == -100")
+          expect(stdout.string).to match "CHECK: INST HEALTH_STATUS TEMP1 == -100 success"
+          stdout.rewind
+
+          check("INST","HEALTH_STATUS","TEMP1","== -100")
+          expect(stdout.string).to match "CHECK: INST HEALTH_STATUS TEMP1 == -100 success"
+          stdout.rewind
+
+          check_formatted("INST HEALTH_STATUS TEMP1 == '-100.000'")
+          expect(stdout.string).to match "CHECK: INST HEALTH_STATUS TEMP1 == '-100.000' success"
+          stdout.rewind
+
+          check_formatted("INST","HEALTH_STATUS","TEMP1","== '-100.000'")
+          expect(stdout.string).to match "CHECK: INST HEALTH_STATUS TEMP1 == '-100.000' success"
+          stdout.rewind
+
+          check_with_units("INST HEALTH_STATUS TEMP1 == '-100.000 C'")
+          expect(stdout.string).to match "CHECK: INST HEALTH_STATUS TEMP1 == '-100.000 C' success"
+          stdout.rewind
+
+          check_with_units("INST","HEALTH_STATUS","TEMP1","== '-100.000 C'")
+          expect(stdout.string).to match "CHECK: INST HEALTH_STATUS TEMP1 == '-100.000 C' success"
+          stdout.rewind
+
+          check_raw("INST HEALTH_STATUS TEMP1")
+          expect(stdout.string).to match "CHECK: INST HEALTH_STATUS TEMP1 == 0"
+        end
+
+        check("INST HEALTH_STATUS TEMP1 < 0")
+        expect { check("INST HEALTH_STATUS TEMP1 > 0") }.to raise_error(Cosmos::CheckError)
+      end
+    end
+
+    describe "check_tolerance, check_tolerance_raw" do
+      it "checks a telemetry item vs tolerance" do
+        capture_io do |stdout|
+          check_tolerance("INST HEALTH_STATUS TEMP1", -100.0, 1)
+          expect(stdout.string).to match "CHECK: INST HEALTH_STATUS TEMP1 was within range"
+          stdout.rewind
+
+          expect { check_tolerance("INST HEALTH_STATUS TEMP1", -200.0, 1) }.to raise_error(CheckError, /CHECK: INST HEALTH_STATUS TEMP1 failed to be within range/)
+          stdout.rewind
+
+          check_tolerance_raw("INST HEALTH_STATUS TEMP1", 0, 1)
+          expect(stdout.string).to match "CHECK: INST HEALTH_STATUS TEMP1 was within range"
+          stdout.rewind
+
+          expect { check_tolerance_raw("INST HEALTH_STATUS TEMP1", 100, 1) }.to raise_error(CheckError, /CHECK: INST HEALTH_STATUS TEMP1 failed to be within range/)
+          stdout.rewind
+        end
+      end
+    end
+
+    describe "check_expression" do
+      it "checks an arbitrary expression" do
+        capture_io do |stdout|
+          check_expression("true == true")
+          expect(stdout.string).to match "CHECK: true == true is TRUE"
+      end
+
+        expect { check_expression("true == false") }.to raise_error(CheckError, "CHECK: true == false is FALSE")
+      end
+    end
+
+    describe "wait, wait_raw, wait_tolerance, wait_tolerance_raw" do
+      it "waits for telemetry check to be true" do
+        capture_io do |stdout|
+          # Success
+          wait("INST HEALTH_STATUS TEMP1 == -100.0", 5)
+          expect(stdout.string).to match "WAIT: INST HEALTH_STATUS TEMP1 == -100.0"
+          stdout.rewind
+          wait_raw("INST HEALTH_STATUS TEMP1 == 0", 5)
+          expect(stdout.string).to match "WAIT: INST HEALTH_STATUS TEMP1 == 0"
+          stdout.rewind
+          wait_tolerance("INST HEALTH_STATUS TEMP1", -100.0, 1, 5)
+          expect(stdout.string).to match "WAIT: INST HEALTH_STATUS TEMP1 was within"
+          stdout.rewind
+          wait_tolerance_raw("INST HEALTH_STATUS TEMP1", 0, 1, 5)
+          expect(stdout.string).to match "WAIT: INST HEALTH_STATUS TEMP1 was within"
+          stdout.rewind
+
+          # Failure
+          wait("INST HEALTH_STATUS TEMP1 == -200.0", 0.1)
+          expect(stdout.string).to match "WAIT: INST HEALTH_STATUS TEMP1 == -200.0 failed"
+          stdout.rewind
+          wait_raw("INST HEALTH_STATUS TEMP1 == 100", 0.1)
+          expect(stdout.string).to match "WAIT: INST HEALTH_STATUS TEMP1 == 100 failed"
+          stdout.rewind
+          wait_tolerance("INST HEALTH_STATUS TEMP1", -200.0, 1, 0.1)
+          expect(stdout.string).to match "WAIT: INST HEALTH_STATUS TEMP1 failed to be within"
+          stdout.rewind
+          wait_tolerance_raw("INST HEALTH_STATUS TEMP1", 100, 1, 0.1)
+          expect(stdout.string).to match "WAIT: INST HEALTH_STATUS TEMP1 failed to be within"
+          stdout.rewind
+        end
+      end
+    end
+
+    describe "wait_expression" do
+      it "waits for an expression to be true" do
+        capture_io do |stdout|
+          # Success
+          wait_expression("true == true", 5)
+          expect(stdout.string).to match "WAIT: true == true is TRUE"
+          stdout.rewind
+
+          # Failure
+          wait_expression("true == false", 0.1)
+          expect(stdout.string).to match "WAIT: true == false is FALSE"
+          stdout.rewind
+        end
+      end
+    end
+
+    describe "wait_check, wait_check_raw, wait_check_tolerance, wait_check_tolerance_raw" do
+      it "waits for telemetry check to be true" do
+        capture_io do |stdout|
+          # Success
+          wait_check("INST HEALTH_STATUS TEMP1 == -100.0", 5)
+          expect(stdout.string).to match "CHECK: INST HEALTH_STATUS TEMP1 == -100.0"
+          stdout.rewind
+          wait_check_raw("INST HEALTH_STATUS TEMP1 == 0", 5)
+          expect(stdout.string).to match "CHECK: INST HEALTH_STATUS TEMP1 == 0"
+          stdout.rewind
+          wait_check_tolerance("INST HEALTH_STATUS TEMP1", -100.0, 1, 5)
+          expect(stdout.string).to match "CHECK: INST HEALTH_STATUS TEMP1 was within"
+          stdout.rewind
+          wait_check_tolerance_raw("INST HEALTH_STATUS TEMP1", 0, 1, 5)
+          expect(stdout.string).to match "CHECK: INST HEALTH_STATUS TEMP1 was within"
+          stdout.rewind
+        end
+
+        # Failure
+        expect { wait_check("INST HEALTH_STATUS TEMP1 == -200.0", 0.1) }.to raise_error(CheckError, /CHECK: INST HEALTH_STATUS TEMP1 == -200.0 failed/)
+        expect { wait_check_raw("INST HEALTH_STATUS TEMP1 == 100", 0.1) }.to raise_error(CheckError, /CHECK: INST HEALTH_STATUS TEMP1 == 100 failed/)
+        expect { wait_check_tolerance("INST HEALTH_STATUS TEMP1", -200.0, 1, 0.1) }.to raise_error(CheckError, /CHECK: INST HEALTH_STATUS TEMP1 failed to be within/)
+
+        expect { wait_check_tolerance_raw("INST HEALTH_STATUS TEMP1", 100, 1, 0.1) }.to raise_error(CheckError, /CHECK: INST HEALTH_STATUS TEMP1 failed to be within/)
+
+      end
+    end
+
+    describe "wait_check_expression" do
+      it "waits for an expression to be true" do
+        capture_io do |stdout|
+          # Success
+          wait_check_expression("true == true", 5)
+          expect(stdout.string).to match "CHECK: true == true is TRUE"
+          stdout.rewind
+        end
+
+        # Failure
+        expect { wait_check_expression("true == false", 0.1) }.to raise_error(CheckError, /CHECK: true == false is FALSE/)
+      end
+    end
+
+    describe "wait_packet, wait_check_packet" do
+      it "waits for a certain number of packets" do
+        capture_io do |stdout|
+          wait_packet("INST","HEALTH_STATUS",1,0.1)
+        end
+
+        expect { wait_check_packet("INST","HEALTH_STATUS",1,0.1) }.to raise_error(CheckError, /INST HEALTH_STATUS expected to be received 1 times but only received 0 times/)
+      end
+    end
+
+    describe "start" do
+      it "starts a script locally" do
+        class ScriptRunnerFrame; def self.instance; false; end; end
+        start("cosmos.rb")
+      end
+
+      it "starts a script without the .rb extension" do
+        class ScriptRunnerFrame; def self.instance; false; end; end
+        start("cosmos")
+      end
+
+      it "raises an error if the script can't be found" do
+        class ScriptRunnerFrame; def self.instance; false; end; end
+        expect { start("unknown_script.rb") }.to raise_error(RuntimeError)
+      end
+
+      it "starts a script within ScriptRunnerFrame" do
+        class ScriptRunnerFrame
+          @@instrumented_cache = {}
+          def self.instance; true; end
+          def self.instrumented_cache; @@instrumented_cache; end
+          def self.instrumented_cache=(value); @@instrumented_cache = value; end
+          def self.instrument_script(file_text, path, bool); "#"; end
+        end
+        start("cosmos.rb")
+        # This one should use the cached version
+        start("cosmos.rb")
+      end
+    end
+
+  end
+end
+
